requirements.txt
pymongo
pymysql
redis
mysql-connector-python